% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{comparecors.2methods.func}
\alias{comparecors.2methods.func}
\title{Comparing the distribution of correaltions for DEG's and FCcorG's for a specific phenotype}
\usage{
comparecors.2methods.func(DEG.cordata = NULL, FCcorG.cordata = NULL,
  ylab = "")
}
\arguments{
\item{DEG.cordata}{data frame with the correaltion of the genes that are most highly differentially
expressed. Use to select.def.func to compute}

\item{FCcorG}{data frame of correlation for most significant Fold Change genes use select.FCcorG.func}
}
\value{
a boxplot graph with two boxplots comparing the correlation distribution for DEG's and FCcorG for a
specific phenotype
}
\description{
Comparing the distribution of correaltions for DEG's and FCcorG's for a specific phenotype
}
