% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{cor.fn1}
\alias{cor.fn1}
\title{Function for getting correlations for phenotype of interest FC and gene expression FC}
\usage{
cor.fn1(de = NULL, dp = NULL, thresh = NULL, method = "pearson",
  na.prop = 0.4, p = 1)
}
\arguments{
\item{de}{= computed gene expression fold changes for every strain}

\item{dp}{= computed phenotype fold changes for every strain}

\item{thresh}{= correlation threshold}

\item{method}{is the correlation method. Here we use pearson correlation}

\item{p}{is the p value threshold, only data with p value smaller than p are computed}
}
\value{
a x by 4 data frame with bicor values and p values for all x number of genes.
Specifies probe, phenotype, bicor, and p value
}
\description{
Need to compute dPhno for a specific phenotype prior
}
