% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{deg.analysis}
\alias{deg.analysis}
\title{function for DEG analysis}
\usage{
deg.analysis(dat_Null = NULL, dat_Alt = NULL, null = NULL,
  alt = NULL)
}
\description{
The function uses limma to calculate he values of differentially expressed genes in two data sets
}
