% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{differences.func}
\alias{differences.func}
\title{function to compute the Fold-Change in gene expression and phenotype between M and F}
\usage{
differences.func(expr.mean.F.unique = NULL, expr.mean.M.unique = NULL,
  pheno.mean.F = NULL, pheno.mean.M = NULL)
}
\arguments{
\item{expr.mean.F.unique}{mean Female expression data for every strain}

\item{expr.mean.M.unique}{mean Male expression data for every strain}

\item{pheno.mean.F}{mean Female phenotype data for every strain}

\item{pheno.mean.M}{mean Male phenotype data for every strain}
}
\value{
a list containing 3 data frames
dExpr contains the FC of every gene expression for every strain
dPhno contains the FC of every phenotype for every strain
dPhno_log is the log of dPhno values
}
\description{
(a substraction because the values are log transformed hence log(A)-log(B) = log (A/B))
Note that the function substracts the Male data from the Female. Hence a positive value means that the Female expression
or phenotype value is greater
!dPhno_log is unnecessary
}
