% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{expr.format.func}
\alias{expr.format.func}
\title{Function Sets expression matrix and annotation}
\usage{
expr.format.func(expr0 = NULL)
}
\arguments{
\item{expr0}{which is the output from the read.table function}
}
\description{
Function Sets expression matrix and annotation
}
