% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{find.gene.name.func}
\alias{find.gene.name.func}
\title{Returns name of gene according to a probe number}
\usage{
find.gene.name.func(probe = "")
}
\arguments{
\item{probe.number}{for which you want to know the gene name of}
}
\value{
gene , the gene name
}
\description{
Returns name of gene according to a probe number
}
