% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{get.means}
\alias{get.means}
\title{function to get means and revised annotations}
\usage{
get.means(expr = NULL, anno = NULL, strains = NULL)
}
\arguments{
\item{expr}{data (can be pheno or expression data) specific to one sex}

\item{anno}{sex specific annotation data containing the strain for each mouse number}

\item{strains}{strain names present for a specific sex}
}
\value{
list composed of 2 data frames expr & anno
}
\description{
gets the sex specific mean expression and mean phenotype data for each strain
}
