% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{heatmap.func}
\alias{heatmap.func}
\title{Creates a heatmap of the FC across strains}
\usage{
heatmap.func(FCcorG.cordata = NULL, dExpr = dExpr, PDF.name = "",
  title = "", fc.cut.heat = NULL, cellwidth = 2, cellheight = 5)
}
\arguments{
\item{FCcorG.cordata}{data frame of correlation for most significant Fold Change genes, use select.FCcorG.func}

\item{dExpr}{dExpr data}

\item{title}{desired title of the heatmap}

\item{cellwidth}{desired width of cell. Has to be adjusted depending on number of strains and genes}

\item{cellheigth}{desired height of cell. Has to be adjusted depending on number of strains and genes}
}
\value{
a PDF file with the heatmap

ah , a list  with the indices and other info about the heatmpa
}
\description{
Function that uses the result of select.FCcorG.func to create a heatmap representing the fold change across strains
for all FC genes for a specific phenotype
}
