% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{maxcor.gene.name.func}
\alias{maxcor.gene.name.func}
\title{Search for max absolute correlation and associated gene name}
\usage{
maxcor.gene.name.func(cors1 = NULL)
}
\arguments{
\item{cors1}{data frame with correlation values for all strains. cors1."pheno" is specific for each phenotype}
}
\value{
a row and a gene name
the row corresponds to the information of the gene with the max corr value
gene name is the name of the probe number found prior
}
\description{
function to search for probe and gene name of gene with max absolute correlation with a phenotype
}
