% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{pheno.format.func}
\alias{pheno.format.func}
\title{The function sets phenotype matrix and annotation
Similar function than expr.format.func}
\usage{
pheno.format.func(pheno = NULL)
}
\value{
formatted phenotype data
}
\description{
The function sets phenotype matrix and annotation
Similar function than expr.format.func
}
\examples{
pheno.format.results=pheno.format.func(pheno0)

}
