% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{pheno.heatmap.func}
\alias{pheno.heatmap.func}
\title{Function to make a single column heatmap of the phenotype following specific indices}
\usage{
pheno.heatmap.func(results.heatmap = NULL, dPhno.pheno = NULL,
  PDF.name = "", cellwidth = 10, cellheight = 5)
}
\arguments{
\item{results.heatmap}{the file corresponding to the output of the heatmap.func. It is specific to each phenotype and contains the indices of the strains}

\item{dPhno.pheno}{data frame of the phenotype FC for a specific phenotype across all strains}

\item{PDF.name}{name of the output PDF}

\item{cellwidth}{desired width of cell}

\item{cellheigth}{desired height of cell}
}
\value{
a PDG with the single column heatmap and a file containing indices
}
\description{
Function to make a single column heatmap of the phenotype following specific indices
}
