% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{plot.expr.func}
\alias{plot.expr.func}
\title{Plot histogram of the distribution of a specific gene for Both variables (Male and Female)}
\usage{
plot.expr.func(expr.mean.M.unique = NULL, expr.mean.F.unique = NULL,
  gene.probe = "", gene.name = "")
}
\arguments{
\item{expr.mean.M.unique}{data frame mean Male gene expression values for all genes in dataset}

\item{expr.mean.F.unique}{data frame mean Female gene expression values for all genes in dataset}

\item{gene.probe}{specific probe number for the gene of interest}

\item{gene.name}{name of the gene for the title}
}
\value{
p, a single color coded histogram of the overlapping gene value distribution for both variables
}
\description{
Gives a gene expression vs.frequency histogram and overlays both variable. Ability to campare
distribution of the gene expression for both variables
}
