% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{plot.hist.ExprFC.func}
\alias{plot.hist.ExprFC.func}
\title{Histogram of the strain specific FC of a specific gene}
\usage{
plot.hist.ExprFC.func(gene.probe = "", gene.name = "")
}
\arguments{
\item{gene.probe}{specific probe number for the gene of interest}

\item{gene.name}{name of the gene for the title}
}
\value{
one histogram
histogram showing the distribution of the FC of a desired gene
}
\description{
The function returns a histogram showing the distribution of the strain specific Fold-Change for a specific gene
}
