% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{plot.hist.gene.func}
\alias{plot.hist.gene.func}
\title{Plot  FC geno vs FC pheno scatter plot and histogram for a gene}
\usage{
plot.hist.gene.func(cors1 = NULL, dPhno = NULL, gene.probe = "",
  pheno.of.interest = "", ylab = "")
}
\arguments{
\item{cors1}{a data frame with the correlation values for a specific phenotype}

\item{dPhno}{a data frame with FC values for a specific phenotype for each strain}

\item{gene.probe}{the probe name associated to the gene of interest. Use maxcor.gene.name.func}

\item{pheno.of.interest}{specific name of phenotype of interest}
}
\value{
two plots hist and plot
plot = scatter plot of gene expression FC vs. phenotype FC
hist = gene expression FC histogram
The correlation value of the gene of interest
}
\description{
Function returns two plots 1)gene expression FC vs. phenotype FC for a specific gene and pheno
2 )histogram FC for a specific gene
}
