% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{plot.pheno.func}
\alias{plot.pheno.func}
\title{Plot histogram of phenotype distribution for Both variables (Male and Female)}
\usage{
plot.pheno.func(pheno.mean.M = NULL, pheno.mean.F = NULL,
  pheno.of.interest = "", xlab = "")
}
\arguments{
\item{pheno.mean.M}{data frame mean Male phenotype values for all phenotypes in dataset}

\item{pheno.mean.F}{data frame mean Female phenotype values for all phenotypes in dataset}

\item{pheno.of.interest}{specific name of phenotype of interest}
}
\value{
p, a single histogram of the overlapping phenotype value distribution for both variables
}
\description{
Gives a phenotype vs.frequency histogram and overlays both variable. Ability to campare
distribution of the phenotype for both variables
}
