% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{probe.to.gene.name.func}
\alias{probe.to.gene.name.func}
\title{The function identifies probes that code for identical genes and takes the mean gene expression value across probes}
\usage{
probe.to.gene.name.func(probe.vs.gene = expr.formatted.gene.annotation,
  expr.mean = NULL)
}
\arguments{
\item{probe.vs.gene}{two column data frame containing the gene name assigned to each probe number}

\item{expr.mean}{data frame containing gene expression data for each strain of interests for a specific sex}
}
\value{
a newly formatted data frame containing gene epression data for each strain of interest

ind.test.4 with only GENE NAMES for a unique gene (no duplicate genes). The gene names are trownames
}
\description{
The first probe in the list is kept an contains reformatted, mean gene expression values
}
