% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{remove.na.func}
\alias{remove.na.func}
\title{Function to filter NA values from phenotype data}
\usage{
remove.na.func(pheno = NULL, na.prop = 0.3)
}
\arguments{
\item{pheno}{pheno.formatted data frame, output from pheno.format.func$pheno}

\item{na.prop}{threshold value}
}
\description{
Function to filter NA values from phenotype data
}
\examples{
pheno6=remove.na.func(pheno=pheno.formatted, na.prop=0.3)
}
