% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{select.deg.func}
\alias{select.deg.func}
\title{Function to select strong differentially expressed genes (DEG's)}
\usage{
\method{select}{deg.func}(data.deg = NULL, minlogFC.thresh = NULL,
  maxpvalue = NULL)
}
\arguments{
\item{data.deg}{the output from deg.analysis function}

\item{minlogFC.thresh}{let's use 1}

\item{minpvalue}{generally 0.05}
}
\value{
selected.deg a data frame with the DEG genes that follow the criterias set
}
\description{
The function uses the output data from the prior executed deg.analysis function to select the genes with
significant DEG, above statistical significance. Threshholds such as the minimum Fold-CHange or a maximum
p-value can be set tot filter the data
}
