% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{select.FCcorG.func}
\alias{select.FCcorG.func}
\title{Function to select genes with highest correlation}
\usage{
\method{select}{FCcorG.func}(cors1 = NULL, bicor.thresh = NULL)
}
\arguments{
\item{cors1}{data frame containing the bicor and p-values for all genes for a specific phenotype}

\item{bicor.thresh}{minimum absolute bicor value |r| for consideration}
}
\value{
cors1.FDR a data frame with the bicor and pvalues satisfying the bicor.thresh condition binded
with the FDR values for each gene that satisfied the condition
}
\description{
Uses p.adjust to compute a false discovery rate for genes satisfying a correlation threshhold condition
and locate promising FCcorGenes
}
