% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{strain.spec.gene.func}
\alias{strain.spec.gene.func}
\title{Function to compute a strain specific gene expression Fold-Change barplot}
\usage{
strain.spec.gene.func(dPhno.interest = NULL, gene.probe = "",
  plot.title = "")
}
\arguments{
\item{dPhno.interest}{specific phenotype FC values for every acceptable (no NA's) strains. Use the
output of the plot.hist.phenoFC.func}

\item{gene.probe}{probe number of the gene of interest. Use gene with nax correlation. Refer to maxcor.gene.name.func}

\item{plot.title}{character of desired plot title}
}
\value{
b , a barplot with the FC for every strain for a specific phenotype
}
\description{
The function plots a color annoted barplot representing the FC for every strain for a specific gene.
Red represents a positive FC meaning the Female gene expression value is greater, Blue is the opposite
The order of the strain is the same as for the strain.spec.pheno.dec.func for the same dPhno.interest and gene
}
