% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{strain.spec.pheno.dec.func}
\alias{strain.spec.pheno.dec.func}
\title{Barplot phenotype FC for every strain for a specific phenotype}
\usage{
strain.spec.pheno.dec.func(dPhno.interest = NULL, gene.probe = NULL,
  plot.title = "")
}
\arguments{
\item{dPhno.interest}{specific phenotype FC values for every acceptable (no NA's) strains. Use the
output of the plot.hist.phenoFC.func)}

\item{gene.probe}{probe number of the gene of interest. Use gene with nax correlation. Refer to maxcor.gene.name.func}

\item{plot.title}{character of desired plot title}
}
\value{
b , a barplot with the FC for every strain for a specific phenotype
}
\description{
The function plots a color annoted barplot representing the FC for every strain for a specific phenotype.
Red represents a positive FC meaning the Female phenotype value is greater, Blue is the opposite
}
