% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_Data.formatting&analysis.functions.R
\name{strains.for.analysis.func}
\alias{strains.for.analysis.func}
\title{selects strains for analysis}
\usage{
strains.for.analysis.func(expr2 = NULL, pheno2 = NULL,
  expr2.samp.annotation = NULL, pheno2.samp.annotation = NULL)
}
\description{
selects strains for analysis
}
\examples{
strains.for.analysis.results=strains.for.analysis.func(expr2=expr.formatted, pheno2=t(pheno6), expr2.samp.annotation=expr.formatted.samp.annotation, pheno2.samp.annotation=pheno.formatted.samp.annotation)

}
