% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_corr_graphs.function.R
\name{venn.diagram.func}
\alias{venn.diagram.func}
\title{Function to create a Venn diagram between genes included
in the DEG analysis and FCcorG analysis for a specific phenotype}
\usage{
venn.diagram.func(DEG.cordata = NULL, FCcorG.cordata = NULL,
  phenotype.of.interest = "")
}
\arguments{
\item{DEG.cordata}{data frame with the correaltion of the genes that are most highly differentially
expressed. Use to select.def.func to compute}

\item{FCcorG.cordata}{data frame of correlation for most significant Fold Change genes use select.FCcorG.func}

\item{phenotype.of.interest}{name of the phenotype that we are looking at (used for the title of the diagram)}
}
\description{
Function to create a Venn diagram between genes included
in the DEG analysis and FCcorG analysis for a specific phenotype
}
